+++ kernel/src/arch/x86/pic.c
@@ -1,9 +1,9 @@
 #include "types.h"
-#include "arch/x86/pic.h"
-#include "arch/x86/asm_io.h"
+#include "kernel/include/arch/x86/pic.h"
+#include "kernel/include/arch/x86/asm_io.h"
 #include "errors.h"
 
-size_t PIC_init()
+uint32 PIC_init()
 {
     io_outb(PIC1_COMMAND, 0x11);
     io_outb(PIC2_COMMAND, 0x11);
@@ -19,17 +19,16 @@ size_t PIC_init()
     return ERROR_SUCCESS;
 }
 
-size_t PIC_setMask(uint8 a_irqLine)
+uint32 PIC_setMask(
+    uint8 a_irqLine)
 {
     uint16 port;
     uint8 value;
 
-    if (a_irqLine < 8)
-    {
+    if (a_irqLine < 8) {
         port = PIC1_DATA;
     }
-    else
-    {
+    else {
         port = PIC2_DATA;
         a_irqLine -= 8;
     }
@@ -40,17 +39,16 @@ size_t PIC_setMask(uint8 a_irqLine)
     return ERROR_SUCCESS;
 }
 
-size_t PIC_clearMask(uint8 a_irqLine)
+uint32 PIC_clearMask(
+    uint8 a_irqLine)
 {
     uint16 port;
     uint8 value;
 
-    if (a_irqLine < 8)
-    {
+    if (a_irqLine < 8) {
         port = PIC1_DATA;
     }
-    else
-    {
+    else {
         port = PIC2_DATA;
         a_irqLine -= 8;
     }
@@ -61,10 +59,9 @@ size_t PIC_clearMask(uint8 a_irqLine)
     return ERROR_SUCCESS;
 }
 
-size_t PIC_maskUnusedIRQs()
+uint32 PIC_maskUnusedIRQs()
 {
-    for (uint8 i = 2; i < 16; i++)
-    {
+    for (uint8 i = 2; i < 16; i++) {
         PIC_setMask(i);
     }
